+++ built/grammar/TreeNotationCodeMirrorMode.js
@@ -9,7 +9,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const textMateScopeToCodeMirrorStyle_1 = require("./textMateScopeToCodeMirrorStyle");
-// import * as CodeMirrorLib from "codemirror"
 class TreeNotationCodeMirrorMode {
     constructor(name, getProgramConstructorMethod, getProgramCodeMethod, codeMirrorLib = undefined) {
         this._name = name;
@@ -67,7 +66,7 @@ class TreeNotationCodeMirrorMode {
         };
     }
     token(stream, state) {
-        return this._advanceStreamAndGetTokenType(stream, state);
+        return this._advanceStreamAndReturnTokenType(stream, state);
     }
     fromTextAreaWithAutocomplete(area, options) {
         this._originalValue = area.value;
@@ -76,7 +75,9 @@ class TreeNotationCodeMirrorMode {
             mode: this._name,
             tabSize: 1,
             indentUnit: 1,
-            hintOptions: { hint: (cmInstance, option) => this.codeMirrorAutocomplete(cmInstance, option) }
+            hintOptions: {
+                hint: (cmInstance, option) => this.codeMirrorAutocomplete(cmInstance, option)
+            }
         };
         Object.assign(defaultOptions, options);
         this._cmInstance = this._getCodeMirrorLib().fromTextArea(area, defaultOptions);
@@ -89,6 +90,7 @@ class TreeNotationCodeMirrorMode {
         cmInstance.on("keyup", (cm, event) => {
             // https://stackoverflow.com/questions/13744176/codemirror-autocomplete-after-any-keyup
             if (!cm.state.completionActive && !excludedKeys[event.keyCode.toString()])
+                // Todo: get typings for CM autocomplete
                 codeMirrorLib.commands.autocomplete(cm, null, { completeSingle: false });
         });
     }
@@ -97,9 +99,12 @@ class TreeNotationCodeMirrorMode {
     }
     codeMirrorAutocomplete(cmInstance, option) {
         return __awaiter(this, void 0, void 0, function* () {
-            const cursor = cmInstance.getCursor();
+            const cursor = cmInstance.getDoc().getCursor();
             const codeMirrorLib = this._getCodeMirrorLib();
             const result = yield this._getParsedProgram().getAutocompleteResultsAt(cursor.line, cursor.ch);
+            // It seems to be better UX if there's only 1 result, and its the word the user entered, to close autocomplete
+            if (result.matches.length === 1 && result.matches[0].text === result.word)
+                return null;
             return result.matches.length
                 ? {
                     list: result.matches,
@@ -115,11 +120,12 @@ class TreeNotationCodeMirrorMode {
         codeMirrorLib.defineMIME("text/" + this._name, this._name);
         return this;
     }
-    _advanceStreamAndGetTokenType(stream, state) {
-        let next = stream.next();
-        while (typeof next === "string") {
+    _advanceStreamAndReturnTokenType(stream, state) {
+        let nextCharacter = stream.next();
+        const lineNumber = this._getLineNumber(stream, state);
+        while (typeof nextCharacter === "string") {
             const peek = stream.peek();
-            if (next === " ") {
+            if (nextCharacter === " ") {
                 if (peek === undefined || peek === "\n") {
                     stream.skipToEnd(); // advance string to end
                     this._incrementLine(state);
@@ -127,15 +133,19 @@ class TreeNotationCodeMirrorMode {
                 return "bracket";
             }
             if (peek === " ") {
-                state.words.push(stream.current());
-                return this._getWordStyle(state.lineIndex, state.words.length);
+                state.wordIndex++;
+                return this._getWordStyle(lineNumber, state.wordIndex);
             }
-            next = stream.next();
+            nextCharacter = stream.next();
         }
-        state.words.push(stream.current());
-        const style = this._getWordStyle(state.lineIndex, state.words.length);
+        state.wordIndex++;
+        const style = this._getWordStyle(lineNumber, state.wordIndex);
         this._incrementLine(state);
         return style;
+    }
+    _getLineNumber(stream, state) {
+        const num = stream.lineOracle.line + 1; // state.lineIndex
+        return num;
     }
     _getWordStyle(lineIndex, wordIndex) {
         const program = this._getParsedProgram();
@@ -143,18 +153,14 @@ class TreeNotationCodeMirrorMode {
         const highlightScope = program.getWordHighlightScopeAtPosition(lineIndex, wordIndex);
         return program ? textMateScopeToCodeMirrorStyle_1.default(highlightScope.split(".")) : undefined;
     }
+    // todo: remove.
     startState() {
         return {
-            words: [],
-            lineIndex: 1
+            wordIndex: 0
         };
     }
-    blankLine(state) {
-        this._incrementLine(state);
-    }
     _incrementLine(state) {
-        state.words.splice(0, state.words.length);
-        state.lineIndex++;
+        state.wordIndex = 0;
     }
 }
 exports.default = TreeNotationCodeMirrorMode;