+++ built/types.d.ts
@@ -28,6 +28,7 @@ export declare namespace types {
     type word = string;
     type triInt = int;
     type filepath = string;
+    type highlightScope = string;
     type fileExtension = string;
     type sortFn = (nodeA: treeNode, nodeB: treeNode) => triInt;
     type filterFn = (node: treeNode, index: int) => boolean;