+++ tasks/test-all.js
@@ -2,8 +2,9 @@
 
 const jtree = require("../index.js")
 const recursiveReadSync = require("recursive-readdir-sync")
+const quack = require("../tests/quack.js")
 
-require("./runBaseTests.js")
+quack.runTestTree(require("../tests/base.js"))
 
 const allFiles = recursiveReadSync(__dirname + "/../tests/")