+++ tests/quack.js
@@ -14,4 +14,18 @@ quack.quickTest = (name, fn) => {
   })
 }
 
+quack.runTestTree = testTree => {
+  const testsToRun = testTree._runOnly.length
+    ? typeof testTree._runOnly === "string"
+      ? [testTree._runOnly]
+      : testTree._runOnly
+    : Object.keys(testTree).filter(key => !key.startsWith("_"))
+  testsToRun.forEach(key => {
+    tap.test(key, function(childTest) {
+      const testCase = testTree[key](childTest.equal)
+      childTest.end()
+    })
+  })
+}
+
 module.exports = quack